Fixed map and commented out unused parts of the page
Part of the map script was missing and the css was a bit weird. Now the map appears below the fold and it takes up the full width and 80% of the height of the screen (so all of it can be on screen at once.)

We could change this later, but now the map is working so we can continue developing around it.

ps: I also removed the txt file that has the batch commands for the local server .bat I made earlier.