Location search box
Location search box is supposed to have a drop down that has suggestions for you. Then when you hit the 'Enter' key it will pinpoint that place on the map. Copy the Java script code into JSFiddle.com to see what it is supposed to do. Thanks guys. I will continue to work on it this weekend.