Place Search Is Up and Running
I got the places search to work now. You have to press the enter key for it to work, therefore we need to make it where we can click the search button instead to initiate the search. We also need to figure out a way to integrate the other boxes into the search and not just the places text form